Move workflow back into tasks
