+++ src/databricks/labs/ucx/hive_metastore/workflows.py
@@ -1,7 +1,6 @@
 from databricks.labs.ucx.assessment.workflows import Assessment
 from databricks.labs.ucx.contexts.workflow_task import RuntimeContext
-from databricks.labs.ucx.framework.tasks import job_task
-from databricks.labs.ucx.framework.workflows import Workflow
+from databricks.labs.ucx.framework.tasks import job_task, Workflow
 from databricks.labs.ucx.hive_metastore.tables import AclMigrationWhat, What
